Break up canvas store selectors by function
