+++ src/renderer/src/features/machine/hooks/useJobActions.ts
@@ -2,7 +2,7 @@ import { useState, useRef } from "react";
 import { v4 as uuid } from "uuid";
 import { useShallow } from "zustand/react/shallow";
 import { useCanvasStore } from "../../../store/canvasStore";
-import { selectJobActionsCanvasState } from "../../../store/canvasStoreSelectors";
+import { selectJobActionsCanvasState } from "../../../store/canvasSelectors";
 import { useTaskStore } from "../../../store/taskStore";
 import { useMachineStore } from "../../../store/machineStore";
 import { parseGcode } from "../../../utils/gcodeParser";
